Created a report.model package as the report package was getting a bit polluted. Renamed "premiumCompaniesTags" to be generic.
