+++ src/main/java/com/ft/asanaapi/AsanaClient.java
@@ -4,8 +4,8 @@ import com.ft.asanaapi.auth.BasicAuthRequestInterceptor;
 import com.ft.asanaapi.auth.NonAsanaUserException;
 import com.ft.asanaapi.model.*;
 import com.ft.config.Config;
-import com.ft.report.ReportTask;
-import com.ft.report.ReportTasksData;
+import com.ft.report.model.ReportTask;
+import com.ft.report.model.ReportTasksData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import retrofit.RestAdapter;
