+++ src/main/java/com/ft/report/model/ReportType.java
@@ -1,4 +1,4 @@
-package com.ft.report;
+package com.ft.report.model;
 
 public enum ReportType {
     SUNDAY_FOR_MONDAY,
