+++ src/main/java/com/ft/services/AsanaService.java
@@ -2,7 +2,7 @@ package com.ft.services;
 
 import com.ft.asanaapi.AsanaClient;
 import com.ft.config.Config;
-import com.ft.report.ReportTask;
+import com.ft.report.model.ReportTask;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.stereotype.Service;
